requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.46
asyncpg==0.31.0
alembic==1.18.3
pydantic==2.12.5
pydantic-settings==2.12.0
python-jose==3.5.0
passlib==1.7.4
argon2-cffi==25.1.0
httpx==0.28.1
Jinja2==3.1.6
python-multipart==0.0.22
prometheus_client==0.22.1
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
aiosqlite==0.22.1
pymupdf
python-docx